added custom exception for wrong id

+++ src/main/java/fr/campus/loic/service/SensorService.java
@@ -1,6 +1,6 @@
 package fr.campus.loic.service;
 
-import fr.campus.loic.interfaces.HeartbeatSensor;
+import fr.campus.loic.exceptions.SensorNotExists;
 import fr.campus.loic.interfaces.SensorDevice;
 import fr.campus.loic.model.Sensor;
 import fr.campus.loic.repository.SensorRepository;
@@ -21,7 +21,10 @@ public class SensorService implements SensorDevice {
         return sensorRepository.getAllSensors();
     }
 
-    public Sensor getSensorById(int id) {
+    public Sensor getSensorById(int id) throws SensorNotExists {
+        if (sensorRepository.getSensor(id) == null){
+            throw new SensorNotExists("Sensor with id " + id + " does not exist");
+        }
         return sensorRepository.getSensor(id);
     }
 
@@ -29,11 +32,17 @@ public class SensorService implements SensorDevice {
         sensorRepository.addSensor(sensor);
     }
 
-    public void removeSensor(int id) {
+    public void removeSensor(int id) throws SensorNotExists {
+        if (sensorRepository.getSensor(id) == null){
+            throw new SensorNotExists("Sensor with id " + id + " does not exist");
+        }
         sensorRepository.removeSensor(id);
     }
 
-    public void modifySensor(int id, Sensor sensor) {
+    public void modifySensor(int id, Sensor sensor) throws SensorNotExists {
+        if (sensorRepository.getSensor(id) == null){
+            throw new SensorNotExists("Sensor with id " + id + " does not exist");
+        }
         Sensor existing = getSensorById(id);
         existing.setName(sensor.getName());
         sensorRepository.addSensor(existing);
